refactor(routing): dig as default page
